refactor(auth): move auth logic to common/route.rs
- Moved `Claims`, `LoginRequest`, `LoginResponse` structs from `common/src/opts.rs` to `common/src/route.rs`.
- Moved `impl RunOpt` block containing auth methods (`create_token`, `check_auth`, `login`) to `common/src/route.rs`.
- Updated `worker/src/lib.rs` imports to reflect these changes.
- Fixed a build error in `native/src/opts.rs` by correctly initializing new auth fields in `RunOpt`.
- Verified compilation for all workspace members (`common`, `worker`, `native`).
