+++ worker/src/lib.rs
@@ -6,7 +6,8 @@ use crate::{ai::WasmAI, d1::WasmD1};
 use frankenstein::{client_reqwest, updates::Update};
 use hjcommon::ai::OpenAI;
 use hjcommon::d1::DB;
-use hjcommon::opts::{LoginRequest, RunOpt};
+use hjcommon::opts::RunOpt;
+use hjcommon::route::LoginRequest;
 use hjcommon::tg::{self, send_random_word};
 use log::{debug, error, info};
 use std::sync::Once;
